chore: improvements on Archon witness population algorithm
* Split the `witness.rs` file into a `populate.rs` file, containing
only the auto population algorithm

* Automatically pad data for committed columns to reach the target
height specified in the `ModuleMode` and the oracle's relative
height

* Fix the calculation for the length of linear combinations

* Add extra comments in the code
